+++ src/component/activity_list.rs
