+++ src/component/activity_view.rs
